+++ auth-client/src/main/java/com/jinelei/iotgenius/auth/client/controller/RoleController.java
@@ -5,10 +5,12 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.PageDTO;
 import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
 import com.github.xiaoymin.knife4j.annotations.ApiSupport;
 import com.jinelei.iotgenius.auth.api.role.RoleApi;
+import com.jinelei.iotgenius.auth.client.configuration.permission.instance.PermissionInstance;
 import com.jinelei.iotgenius.auth.client.domain.RoleEntity;
 import com.jinelei.iotgenius.auth.client.helper.PageHelper;
 import com.jinelei.iotgenius.auth.client.service.RoleService;
 import com.jinelei.iotgenius.auth.dto.role.*;
+import com.jinelei.iotgenius.auth.helper.AuthorizationHelper;
 import com.jinelei.iotgenius.auth.permission.declaration.RoleDeclaration;
 import com.jinelei.iotgenius.common.request.PageRequest;
 import com.jinelei.iotgenius.common.view.BaseView;
@@ -45,6 +47,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "创建角色")
     @PostMapping("/create")
     public BaseView<Void> create(@RequestBody @Valid RoleCreateRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_CREATE);
         roleService.create(request);
         return new BaseView<>("创建成功");
     }
@@ -54,6 +57,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "删除角色")
     @PostMapping("/delete")
     public BaseView<Void> delete(@RequestBody @Valid RoleDeleteRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_DELETE);
         roleService.delete(request);
         return new BaseView<>("删除成功");
     }
@@ -63,6 +67,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "更新角色")
     @PostMapping("/update")
     public BaseView<Void> update(@RequestBody @Valid RoleUpdateRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_UPDATE);
         roleService.update(request);
         return new BaseView<>("更新成功");
     }
@@ -72,6 +77,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "获取角色")
     @PostMapping("/get")
     public BaseView<RoleResponse> get(@RequestBody @Valid RoleQueryRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_DETAIL);
         RoleEntity entity = roleService.get(request);
         RoleResponse convert = roleService.convert(entity);
         return new BaseView<>(convert);
@@ -82,6 +88,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "获取角色树")
     @PostMapping("/tree")
     public BaseView<List<RoleResponse>> tree(@RequestBody @Valid RoleQueryRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_SUMMARY);
         List<RoleEntity> entities = roleService.tree(request);
         List<RoleResponse> convert = roleService.convertTree(entities);
         return new BaseView<>(convert);
@@ -92,6 +99,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "获取角色列表")
     @PostMapping("/list")
     public ListView<RoleResponse> list(@RequestBody @Valid RoleQueryRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_SUMMARY);
         List<RoleEntity> entities = roleService.list(request);
         List<RoleResponse> convert = entities.parallelStream().map(entity -> roleService.convert(entity))
                 .collect(Collectors.toList());
@@ -103,6 +111,7 @@ public class RoleController implements RoleApi {
     @Operation(summary = "获取角色分页")
     @PostMapping("/page")
     public PageView<RoleResponse> page(@RequestBody @Valid PageRequest<RoleQueryRequest> request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.ROLE_SUMMARY);
         IPage<RoleEntity> page = roleService.page(PageHelper.toPage(new PageDTO<>(), request), request.getParams());
         List<RoleResponse> collect = page.getRecords().parallelStream().map(entity -> roleService.convert(entity))
                 .collect(Collectors.toList());
