+++ auth-client/src/main/java/com/jinelei/iotgenius/auth/client/controller/UserController.java
@@ -5,10 +5,12 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.PageDTO;
 import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
 import com.github.xiaoymin.knife4j.annotations.ApiSupport;
 import com.jinelei.iotgenius.auth.api.user.UserApi;
+import com.jinelei.iotgenius.auth.client.configuration.permission.instance.PermissionInstance;
 import com.jinelei.iotgenius.auth.client.domain.UserEntity;
 import com.jinelei.iotgenius.auth.client.helper.PageHelper;
 import com.jinelei.iotgenius.auth.client.service.UserService;
 import com.jinelei.iotgenius.auth.dto.user.*;
+import com.jinelei.iotgenius.auth.helper.AuthorizationHelper;
 import com.jinelei.iotgenius.common.request.PageRequest;
 import com.jinelei.iotgenius.common.view.BaseView;
 import com.jinelei.iotgenius.common.view.ListView;
@@ -42,6 +44,7 @@ public class UserController implements UserApi {
     @Operation(summary = "创建用户")
     @PostMapping("/create")
     public BaseView<Void> create(@RequestBody @Valid UserCreateRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_CREATE);
         userService.create(request);
         return new BaseView<>("创建成功");
     }
@@ -51,6 +54,7 @@ public class UserController implements UserApi {
     @Operation(summary = "删除用户")
     @PostMapping("/delete")
     public BaseView<Void> delete(@RequestBody @Valid UserDeleteRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_DELETE);
         userService.delete(request);
         return new BaseView<>("删除成功");
     }
@@ -60,6 +64,7 @@ public class UserController implements UserApi {
     @Operation(summary = "更新用户")
     @PostMapping("/update")
     public BaseView<Void> update(@RequestBody @Valid UserUpdateRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_UPDATE);
         userService.update(request);
         return new BaseView<>("更新成功");
     }
@@ -69,6 +74,7 @@ public class UserController implements UserApi {
     @Operation(summary = "获取用户")
     @PostMapping("/get")
     public BaseView<UserResponse> get(@RequestBody @Valid UserQueryRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_DETAIL);
         UserEntity entity = userService.get(request);
         UserResponse convert = userService.convert(entity);
         return new BaseView<>(convert);
@@ -79,6 +85,7 @@ public class UserController implements UserApi {
     @Operation(summary = "获取用户列表")
     @PostMapping("/list")
     public ListView<UserResponse> list(@RequestBody @Valid UserQueryRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_SUMMARY);
         List<UserEntity> entities = userService.list(request);
         List<UserResponse> convert = entities.parallelStream().map(entity -> userService.convert(entity))
                 .collect(Collectors.toList());
@@ -90,6 +97,7 @@ public class UserController implements UserApi {
     @Operation(summary = "获取用户分页")
     @PostMapping("/page")
     public PageView<UserResponse> page(@RequestBody @Valid PageRequest<UserQueryRequest> request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_SUMMARY);
         IPage<UserEntity> page = userService.page(PageHelper.toPage(new PageDTO<>(), request), request.getParams());
         List<UserResponse> collect = page.getRecords().parallelStream().map(entity -> userService.convert(entity))
                 .collect(Collectors.toList());
@@ -119,6 +127,7 @@ public class UserController implements UserApi {
     @Operation(summary = "用户修改密码")
     @PostMapping("/updatePassword")
     public BaseView<String> updatePassword(UserUpdatePasswordRequest request) {
+        AuthorizationHelper.checkPermission(PermissionInstance.USER_UPDATE);
         userService.updatePassword(request);
         return new BaseView<>("修改成功");
     }
